backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
httpx==0.27.2
gitpython==3.1.43
langgraph
langchain-core
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-groq
mcp
python-multipart==0.0.20
python-dotenv==1.0.1
click==8.1.7
pytest==8.3.4
pytest-asyncio==0.24.0
anyio==4.7.0
